fix filename f32.c

+++ core/type/f32.h
@@ -3,7 +3,7 @@
 
 #include <self.h>
 
-float *f32New(int x);
+float *f32New(float x);
 Self   f32Self();
 
 #endif  // F32_H
